Move async utilities to their own file
